+++ on_off_controller/on_off_controller.ino
@@ -1,5 +1,7 @@
 #include "connect_funcs.h"
 
+#include "mqtt_funcs.h"
+
 void setup() {
   Serial.begin(115200);
   Serial.print("\nInit ");
@@ -11,12 +13,13 @@ void setup() {
   randomSeed(analogRead(0));
   pinMode(LED_BUILTIN, OUTPUT);
   setup_wifi();
-  client.setServer(MQTT_SERVER, MQTT_PORT);
-  client.setCallback(handleMqttMsgReceived);
+  mqtt_client.setServer(MQTT_SERVER, MQTT_PORT);
+  mqtt_client.setCallback(handleMqttMsgReceived);
   // pinMode(ON_OR_OFF_PIN, OUTPUT);
 }
 
 void loop() {
-  if (!client.connected()) connect_broker();
-  client.loop();
+  if (!mqtt_client.connected())
+    connect_mqtt_broker(mqtt_client, publishEspStateOnOrOff);
+  mqtt_client.loop();
 }
